Bretonnia spellcasting trainer (Morgana le Fay)
The Fey Enchantress is the only spell teacher
There exist multiple copies of it ( the encyclopedia had been deactivated for all Trainers, so nobody will ever notice)
It is thought that the fey enchantress ports around and is magically just in the next castle... how convienient.

The fey chantress can teach damsels lore of heavens once it is unlocked.

The fey enchantress will not teach the player, unless a prophetess is part of the party or the player itself is a prophetess.

Careers have been excluded from learning magic, unless special conditions are met.

The fey enchantress will sell scrolls to anybody.