reenabled seeker parameter
since the new UI doesnt allow to select targets during the ability selection, the bug for scrolling is gone for good. time to test again if all works fine.

Renabled for Dart the target seeking

Also added for Shadow blood of the Necromancer target seeking (discussed with a lore guy)