adjusted darkenergy costs according r4mpzy discussion
all VC undead troops  (beside wraiths and vampires) are not having an upkeep, but a high upgrade cost.

upgrade money costs are removed for all undead ( not vampires)

for every causalty that is not a prisoner you get 1 dark energy.

Vampire and wraiths have a high amount of Dark energy upkeep cost.