zoologist and the imperial menagery
no scene or anything, just a zoologist in the keep that offers for a huge amount of prestige a demi gryph.

choice is prompted...

known issues:
Pictures of the gryphs are totally off ( better remove)
Names are not changed and its impossible to take right one

heavily needs some revision by lore. Mechanic is discussed, not the implementation