necromancy fix
we decided that companions that have the ability to raise dead

Can raise dead is only used on player, so if the player is sylvanian or mousillon it should have the first check passed.

if a necromancer (who has the lore necromancy ) is present or attribute, he should be able to peform raise dead.