Compile for 1.3.4 + introduce custom voice manager
Custom voice manager is an alternative way to play sounds because currently the base game implementation of Agent.MakeVoice -> SkinVoiceType crashes in the engine. This is not meant for release, only to be able to test sounds even while we are waiting for a fix. Turn it on in config if you want, it is off by default.