fix blessings from shrines
the shrines were adding way to much devotion for religions. this was due to a bug in the addition calculation
also once fanatic was reached, the change amount did not allow for changes again back to other religions.

Ui wise, for testing but handy improvement: it displays now all devotions above 0. when cheatmode is active we also display the exact devotion level.